Create Chukwuemeka-Faith-Ogechukwukamma-2024714041
This Python program implements a grading system using 'if' and 'elif'. It takes a student's score as input and prints the corresponding grade based on the following scale:

A 70--100
B 60--69
C 50---59
D 45-49
E 40--44
F 0--39

For each valid score, the program also prints 'YES' to confirm the grading was successful. If the score was outside the valid range, the program prints 'INVALID SCORE' or 'NO'

Created by: Chukwuemeka Faith Ogechukwu 
Reg No: 2024714041